* Refactor Ecpay and separate module Helper and Notification from Ecpay   into two files * Implement validation methods for notifications
