+++ lib/offsite_payments/integrations/ecpay.rb
@@ -65,58 +65,9 @@ module OffsitePayments #:nodoc:
         ]
       end
 
-      class Helper < OffsitePayments::Helper
-        def initialize(order, account, options = {})
-          super
-        end
-
-        Ecpay.support_keys.each do |parameter|
-          # TODO: Use pure Ruby to implement #underscore
-          mapping parameter.underscore.to_sym, parameter
-        end
-
-        mapping :amount, 'TotalAmount'
-
-        def form_fields
-          sign_fields
-        end
-
-        def sign_fields
-          @fields.merge!('CheckMacValue' => generate_signature)
-        end
-
-        def messages
-          @messages = @fields.slice(*Ecpay.support_keys).sort
-          @messages.unshift(['HashKey', Ecpay.hash_key])
-          @messages.push(['HashIV', Ecpay.hash_iv])
-        end
-
-        def generate_signature
-          @fields['MerchantID'] ||= Ecpay.merchant_id
-          @fields['EncryptType'] ||= 1
-
-          query = messages.map { |key, value| [key, value].join('=') } * '&'
-          encoded_url = CGI.escape(query).downcase
-          Digest::SHA256.hexdigest(encoded_url).upcase
-        end
-      end
-
-      class Notification < OffsitePayments::Notification
-        def initialize(post, options = {})
-          super
-        end
-
-        Ecpay.support_keys.each do |parameter|
-          # TODO: Use pure Ruby to implement #underscore
-          define_method(parameter.underscore) do
-            params[parameter]
-          end
-        end
-
-        def merchant_id
-          @params['MerchantID']
-        end
-      end
+      require 'offsite_payments/integrations/ecpay/concern/has_trade_info'
+      require 'offsite_payments/integrations/ecpay/helper'
+      require 'offsite_payments/integrations/ecpay/notification'
 
       def self.setup
         yield(self)
