+++ lib/offsite_payments/integrations/ecpay/notification.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+module OffsitePayments
+  module Integrations
+    module Ecpay
+      class Notification < OffsitePayments::Notification
+        include HasTradeInfo
+
+        def initialize(post, options = {})
+          super
+        end
+
+        Ecpay.support_keys.each do |parameter|
+          # TODO: Use pure Ruby to implement #underscore
+          define_method(parameter.underscore) do
+            params[parameter]
+          end
+        end
+
+        def valid?
+          verify_mac trade_info
+        end
+
+        def succeed?
+          @params['RtnCode'].to_i == 1
+        end
+
+        def merchant_id
+          @params['MerchantID']
+        end
+
+        def order_id
+          @params['MerchantTradeNo']
+        end
+      end
+    end
+  end
+end
